Refactor: 반차/반반차 사용으로 인한 자료형 변경
소수점 계산을 위해 int에서 double로 변경

+++ BaekEnd/grid/src/main/java/org/highfives/grid/vacation/command/vo/GiveVacation.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class GiveVacation {
-    private int vacationNum;
+    private double vacationNum;
     private String endTime;
     private int employeeId;
     private int typeId;
